requirements.txt
PyMatting==1.1.2
torchvision==0.8.0
tensorflow==2.3.1
torch==1.7.0+cu101
numpy==1.17.0
opencv_python==4.5.1.48
model1==1.0.4.8
Pillow==8.2.0
